refactor: create tool_registry.py
Separate tools from agent.py for modularization
